# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: WaCalls API
  version: "1.0.0"
  description: |
    API HTTP do **WaCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. Todas as rotas aceitam o header **`X-Client-Id`** (qualquer
    identificador estável), usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema: { type: string, example: meuapp }
      description: Identificador do cliente (dono da chamada / filtro de eventos).
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
...
  schemas:
    Session:
      type: object
      properties:
        id: { type: string, example: 6458cca60a17c0cefa71de9da9dc92a2 }
        name: { type: string, example: WhatsApp }
        jid: { type: string, example: "556192264108:13@s.whatsapp.net" }
        state: { type: string, enum: [open, connecting, close], example: open }
        paired: { type: boolean, example: true }
        recording: { type: boolean, example: false, description: "Gravação de chamadas ligada para esta conta" }
...
paths:
...
  /api/sessions:
    get:
      tags: [Sessões]
      summary: Lista as contas
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
    post:
      tags: [Sessões]
      summary: Cria uma conta
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "Minha conta" }
      responses:
        "200":
          description: Criada
          content:
            application/json:
              schema:
                type: object
                properties: { id: { type: string } }
  /api/sessions/{sid}/recording:
    get:
      tags: [Sessões]
      summary: Estado da gravação de chamadas da conta
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
    put:
      tags: [Sessões]
      summary: Liga/desliga a gravação de chamadas da conta
      description: |
        Quando ligado, TODAS as chamadas desta conta são gravadas. No fim de cada
        chamada o MP3 vira uma **nota privada** na conversa do número no Chatwoot
        (se configurado) e um evento `recording` é enviado no webhook da sessão.
        O áudio também fica em `GET /recordings/{callId}.mp3`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties: { enabled: { type: boolean, example: true } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { enabled: { type: boolean } }
  /recordings/{id}:
    get:
      tags: [Sessões]
      summary: Baixa o MP3 de uma gravação (rota pública / capability)
      description: "Fora de /api/ (sem API key). O id (callId) é não-enumerável e atua como capability."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: "abc123.mp3" }
      responses:
        "200":
          description: Áudio MP3
          content:
            audio/mpeg: { schema: { type: string, format: binary } }
        "404": { description: Não encontrado }
  /api/sessions/{sid}:
    delete:
      tags: [Sessões]
      summary: Remove a conta
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Removida }
  /api/sessions/{sid}/pair:
    post:
      tags: [Sessões]
      summary: Inicia o pareamento (QR sai via SSE)
      description: Dispara a geração do QR Code; acompanhe os eventos `session-qr` / `auth-state` em `/api/events`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Pareamento iniciado }
  /api/sessions/{sid}/logout:
    post:
      tags: [Sessões]
      summary: Desconecta a conta (mantém a sessão)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Desconectada }
  /api/sessions/{sid}/calls:
    get:
      tags: [Chamadas]
      summary: Quantas chamadas ativas a sessão tem (e o limite)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  active: { type: integer, example: 2 }
                  maxCallsPerSession: { type: integer, example: 8 }
    post:
      tags: [Chamadas]
      summary: Inicia uma chamada (faz o número de destino tocar)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartCallRequest" }
      responses:
        "200":
          description: Chamada iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  call:
                    type: object
                    properties: { callId: { type: string } }
        "503": { description: Conta não pareada }
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      tags: [Chamadas]
      summary: Atende uma chamada recebida
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200": { description: Atendida }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      tags: [Chamadas]
      summary: Rejeita uma chamada recebida
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200": { description: Rejeitada }
  /api/sessions/{sid}/calls/{id}:
    delete:
      tags: [Chamadas]
      summary: Encerra a chamada
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Encerrada }
  /api/sessions/{sid}/messages/text:
    post:
      tags: [Mensagens]
      summary: Envia uma mensagem de texto
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, text]
              properties:
                to: { type: string, example: "5561999999999", description: "Número (DDI+DDD) ou JID completo (...@s.whatsapp.net / ...@g.us)" }
                text: { type: string, example: "Olá!" }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
        "503": { description: Conta não pareada }
  /api/sessions/{sid}/messages/image:
    post:
      tags: [Mensagens]
      summary: Envia uma imagem
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MediaBody"
                - type: object
                  properties:
                    caption: { type: string, example: "minha foto" }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
  /api/sessions/{sid}/messages/audio:
    post:
      tags: [Mensagens]
      summary: Envia um áudio (use ptt=true para nota de voz)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MediaBody"
                - type: object
                  properties:
                    ptt: { type: boolean, example: true, description: "true = nota de voz (push-to-talk)" }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
  /api/sessions/{sid}/messages/video:
    post:
      tags: [Mensagens]
      summary: Envia um vídeo
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MediaBody"
                - type: object
                  properties:
                    caption: { type: string }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
  /api/sessions/{sid}/messages/document:
    post:
      tags: [Mensagens]
      summary: Envia um documento/arquivo
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/MediaBody"
                - type: object
                  properties:
                    filename: { type: string, example: "contrato.pdf" }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
  /api/sessions/{sid}/webhook:
    get:
      tags: [Webhook]
      summary: Consulta a URL de webhook da sessão
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { webhook: { type: string } }
    post:
      tags: [Webhook]
      summary: Define a URL de webhook (recebe os eventos da sessão)
      description: |
        Eventos enviados (POST JSON) no formato `{ session, event, timestamp, data }`.
        `event` pode ser `message` (mensagem recebida, com `data.text`, `data.type`, `data.raw`...)
        ou `receipt` (confirmações de entrega/leitura).
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, example: "https://meu-sistema/webhook/wacalls" }
      responses:
        "200":
          description: Salvo
          content:
            application/json:
              schema:
                type: object
                properties: { webhook: { type: string } }
    delete:
      tags: [Webhook]
      summary: Remove o webhook da sessão
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Removido }
  /api/sessions/{sid}/chatwoot:
    get:
      tags: [Chatwoot]
      summary: Consulta a integração Chatwoot da sessão (token omitido)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  chatwoot: { type: object }
    post:
      tags: [Chatwoot]
      summary: Conecta a sessão a uma inbox do Chatwoot (canal API)
      description: |
        Configura a integração. Crie no Chatwoot uma inbox do tipo **API**, pegue o
        `inbox_id`, `inbox_identifier`, o `account_id` e um `account_token` (Access Token
        do perfil/agente). Depois aponte o **webhook da inbox** para
        `POST /api/sessions/{sid}/chatwoot/webhook` para as respostas dos agentes saírem no WhatsApp.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, account_id, account_token, inbox_id]
              properties:
                url: { type: string, example: "https://chatwoot.seudominio.com" }
                account_id: { type: integer, example: 1 }
                account_token: { type: string, example: "xxxxxxxxxxxxxxxx" }
                inbox_id: { type: integer, example: 5 }
                inbox_identifier: { type: string, example: "abcd1234" }
                groups: { type: boolean, default: false, description: "Abrir conversas de GRUPOS no Chatwoot (mensagens prefixadas com o autor)" }
                channels: { type: boolean, default: false, description: "Abrir conversas de CANAIS (newsletters) no Chatwoot" }
      responses:
        "200": { description: Conectado }
        "400": { description: Campos obrigatórios faltando }
    delete:
      tags: [Chatwoot]
      summary: Desconecta a integração Chatwoot
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Desconectado }
  /api/sessions/{sid}/chatwoot/groups/{gid}/open:
    post:
      tags: [Chatwoot]
      summary: Abre (sob demanda) uma conversa de GRUPO no Chatwoot
      description: |
        Cria/garante o contato (o próprio grupo) e uma conversa na inbox, sem
        esperar chegar mensagem. Requer Chatwoot configurado na sessão.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/Gid" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contactId: { type: integer }
                  conversationId: { type: integer }
                  chatId: { type: string, example: "120363000000000000@g.us" }
        "400": { description: Chatwoot não configurado ou JID inválido }
  /api/sessions/{sid}/chatwoot/channels/{id}/open:
    post:
      tags: [Chatwoot]
      summary: Abre (sob demanda) uma conversa de CANAL no Chatwoot
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ChannelId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contactId: { type: integer }
                  conversationId: { type: integer }
                  chatId: { type: string, example: "120363000000000000@newsletter" }
        "400": { description: Chatwoot não configurado ou JID inválido }
  /api/sessions/{sid}/chatwoot/webhook:
    post:
      tags: [Chatwoot]
      summary: Endpoint que recebe o webhook do Chatwoot (agente -> WhatsApp)
      description: |
        Configure este endereço como **webhook da inbox** no Chatwoot. Ao receber
        `message_created`/outgoing, envia a mensagem (texto e anexos) pelo WhatsApp.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: "Payload do webhook do Chatwoot" }
      responses:
        "200": { description: Processado }
  /api/sessions/{sid}/history:
    get:
      tags: [Eventos & Histórico]
      summary: Histórico de chamadas da conta (últimas 50)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items: { type: object }
...
  /api/sessions/{sid}/contacts/{jid}/block:
    post:
      tags: [Contatos]
      summary: Bloqueia o contato
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/Jid" }
      responses:
        "200": { description: OK }
  /api/sessions/{sid}/contacts/{jid}/unblock:
    post:
      tags: [Contatos]
      summary: Desbloqueia o contato
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/Jid" }
      responses:
        "200": { description: OK }
...